[Refactor] queryDSL 테스트 관련 config 글로벌로 이동
